--- a/flink-core/src/test/java/org/apache/flink/api/java/tuple/TupleGenerator.java
+++ b/flink-core/src/test/java/org/apache/flink/api/java/tuple/TupleGenerator.java
@@ -135,6 +135,8 @@ class TupleGenerator {
 		StringBuilder sb = new StringBuilder();
 		sb.append("\tpublic static Tuple newInstance(int arity) {\n");
 		sb.append("\t\tswitch (arity) {\n");
+		// special case for Tuple0:
+		sb.append("\t\t\tcase 0: return Tuple0.INSTANCE;\n");
 		for (int i = FIRST; i <= LAST; i++) {
 			sb.append("\t\t\tcase ").append(i).append(": return new Tuple").append(i).append("();\n");
 		}