## What is the purpose of the change

Inside `KeySelectorUtil`, every `ComparableKeySelector#getKey()` call currently creates a new tuple from `Tuple.getTupleClass(keyLength).newInstance();` which seems expensive. Instead, we could get a template tuple and use `Tuple#copy()` which copies the right sub-class in a more optimal way.

Similarly, `ArrayKeySelector` instantiates new Tuple instances via reflection (albeit caching the required tuple class for the returned key) which can be changed the same way.

With the micro-benchmarks added for a very simple job basically only doing a `keyBy()` (https://github.com/dataArtisans/flink-benchmarks/pull/5), I get these results:

```
Benchmark                    Mode  Cnt     Score     Error   Units
------------- old ---------------
KeyByBenchmarks.arrayKeyBy  thrpt    9  1055.706 ± 170.221  ops/ms
KeyByBenchmarks.tupleKeyBy  thrpt    9  1537.923 ± 271.665  ops/ms
------------- new ---------------
KeyByBenchmarks.arrayKeyBy  thrpt    9  1213.073 ±  39.672  ops/ms
KeyByBenchmarks.tupleKeyBy  thrpt    9  1848.172 ± 188.013  ops/ms
```
That is roughly 15% more for the `ArrayKeySelector` and 20% more for the `ComparableKeySelector`.

## Brief change log

- optimise `ComparableKeySelector` using a template `Tuple` instance determined once
- optimise `ArrayKeySelector` using a template `Tuple` instance determined once

## Verifying this change

This change is already covered by existing tests.

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): **no**
  - The public API, i.e., is any changed class annotated with `@Public(Evolving)`: **no**
  - The serializers: **no**
  - The runtime per-record code paths (performance sensitive): **yes**
  - Anything that affects deployment or recovery: JobManager (and its components), Checkpointing, Yarn/Mesos, ZooKeeper: **no**
  - The S3 file system connector: **no**

## Documentation

  - Does this pull request introduce a new feature? **no**
  - If yes, how is the feature documented? **not applicable**