@NicoK 2018-07-10
alright - tried that change which also has the advantage of not using up memory for the `templateKey` - these are the results from the same Benchmark (same hardware, new software though, so I also re-did the old benchmark and ran them twice):
```
old:

Benchmark                    Mode  Cnt     Score    Error   Units
KeyByBenchmarks.arrayKeyBy  thrpt   30  1151.305 ± 21.096  ops/ms
KeyByBenchmarks.arrayKeyBy  thrpt   30  1117.486 ± 43.508  ops/ms
KeyByBenchmarks.tupleKeyBy  thrpt   30  1659.634 ± 28.627  ops/ms
KeyByBenchmarks.tupleKeyBy  thrpt   30  1554.265 ± 82.604  ops/ms

new:

KeyByBenchmarks.arrayKeyBy  thrpt   30  1150.552 ± 51.185  ops/ms
KeyByBenchmarks.arrayKeyBy  thrpt   30  1195.777 ± 10.621  ops/ms
KeyByBenchmarks.tupleKeyBy  thrpt   30  1743.633 ± 27.109  ops/ms
KeyByBenchmarks.tupleKeyBy  thrpt   30  1697.885 ± 22.101  ops/ms
```