@NicoK 2018-07-19
Thanks for the review, merging...